Refactoring code to import data from JSON file
