--- FloridaCounties/Classes/classes.cd
@@ -1,53 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1" MembersFormat="FullSignature">
-  <Class Name="FloridaCounties.Classes.RootObject">
-    <Position X="6.75" Y="0.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Classes\Classes.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="features" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="FloridaCounties.Classes.Feature">
-    <Position X="6.75" Y="2.5" Width="1.5" />
-    <AssociationLine Name="attributes" Type="FloridaCounties.Classes.Attributes" FixedFromPoint="true" FixedToPoint="true">
-      <Path>
-        <Point X="6.75" Y="3.062" />
-        <Point X="5.604" Y="3.062" />
-        <Point X="5.604" Y="4" />
-      </Path>
-    </AssociationLine>
-    <AssociationLine Name="geometry" Type="FloridaCounties.Classes.Geometry" FixedFromPoint="true" FixedToPoint="true">
-      <Path>
-        <Point X="8.25" Y="3.062" />
-        <Point X="9.125" Y="3.062" />
-        <Point X="9.125" Y="4.25" />
-      </Path>
-    </AssociationLine>
-    <TypeIdentifier>
-      <HashCode>AACAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Classes\Classes.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="attributes" />
-      <Property Name="geometry" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="FloridaCounties.Classes.Attributes">
-    <Position X="4.25" Y="4" Width="2.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAEAAAAAAAAAAAAAAAEAAAAAEAAAEAAAAAAA=</HashCode>
-      <FileName>Classes\Classes.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="FloridaCounties.Classes.Geometry">
-    <Position X="8" Y="4.25" Width="2.25" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Classes\Classes.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
