+++ FloridaCounties/DTO/Geometry.cs
@@ -1,25 +1,9 @@
-﻿using NetTopologySuite;
+﻿using System;
+using NetTopologySuite;
 using NetTopologySuite.Geometries;
 using System.Linq;
 
-namespace FloridaCounties.Classes {
-    public class RootObject {
-
-        public Feature[] features { get; set; }
-    }
-
-    public class Feature {
-        public Attributes attributes { get; set; }
-        public Geometry geometry { get; set; }
-    }
-
-    public class Attributes {
-        public int OBJECTID { get; set; }
-        public string COUNTYNAME { get; set; }
-        public int DEPCODE { get; set; }
-        public string COUNTY { get; set; }
-    }
-
+namespace FloridaCounties.Dto {
     public class Geometry {
         private const int DEFAULT_SRID = 4326;
         public float[][][] rings { get; set; }
@@ -28,25 +12,25 @@ namespace FloridaCounties.Classes {
             LinearRing shell = new LinearRing(g.rings[0].Select(p => new Coordinate(p[0], p[1])).ToArray());
             if (!shell.IsCCW) shell = shell.Reverse() as LinearRing;
 
-            if(g.rings.Length > 1) {
-                 LinearRing[] holes = g.rings.Skip(1)
-                    .Select(sp => sp.Select(p => new Coordinate(p[0], p[1])))
-                    .Select(r => new LinearRing(r.ToArray()))
-                    .ToArray();
-                
-                for(int i = 0; i < holes.Length; i++) {
+            if (g.rings.Length > 1) {
+                LinearRing[] holes = g.rings.Skip(1)
+                   .Select(sp => sp.Select(p => new Coordinate(p[0], p[1])))
+                   .Select(r => new LinearRing(r.ToArray()))
+                   .ToArray();
+
+                for (int i = 0; i < holes.Length; i++) {
                     if (holes[i].IsCCW) holes[i] = holes[i].Reverse() as LinearRing;
                 }
 
                 Polygon result = new Polygon(shell, holes) { SRID = DEFAULT_SRID };
                 return result;
             } else {
-                Polygon result = new Polygon(shell) { SRID = DEFAULT_SRID};
+                Polygon result = new Polygon(shell) { SRID = DEFAULT_SRID };
                 return result;
             }
         }
 
-        private static Polygon PolygonFromCoordsArray (float[][] ring) {
+        private static Polygon PolygonFromCoordsArray(float[][] ring) {
             LinearRing shell = new LinearRing(ring.Select(point => new Coordinate(point[0], point[1])).ToArray());
             shell = shell.IsCCW ? shell : (LinearRing)shell.Reverse();
 
@@ -54,7 +38,7 @@ namespace FloridaCounties.Classes {
         }
 
         public static implicit operator MultiPolygon(Geometry g) {
-            MultiPolygon result = new MultiPolygon(g.rings.Select(ring => PolygonFromCoordsArray(ring)).ToArray()) { 
+            MultiPolygon result = new MultiPolygon(g.rings.Select(ring => PolygonFromCoordsArray(ring)).ToArray()) {
                 SRID = DEFAULT_SRID
             };
